requirements.txt
fastapi~=0.109.1
ultralytics~=8.1.16
pillow~=10.2.0
PyYAML~=6.0.1
supervision~=0.18.0
pandas~=2.2.0
numpy~=1.26.3
uvicorn~=0.27.0.post1
pydantic~=2.6.0
lancedb>=0.4.3
aiofiles~=23.2.1